+++ src/WinGet.psm1
@@ -6,7 +6,6 @@ $script:AcceptLicense = "AcceptLicense"
 $script:Force = "Force"
 $script:PackageSource = "WinGet"
 $script:AllVersions = "AllVersions"
-$script:Detailed = "Detailed"
 
 # Utility variables
 # Fast Package References are passed between cmdlets in the format of '<name>#<version>#<source>'